updated to latest PCIDSK SDK source
git-svn-id: https://svn.osgeo.org/gdal/trunk@20074 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/pcidsk/sdk/core/sysvirtualfile.cpp
@@ -88,7 +88,7 @@ SysVirtualFile::SysVirtualFile( CPCIDSKFile *file, int start_block,
         next_block = block_map_data.GetInt( offset + 20, 8 );
     }
 
-    assert( block_index.size() * block_size >= file_length );
+    assert( (uint64)block_index.size() * (uint64)block_size >= file_length );
 }
 
 /************************************************************************/
